+++ common/src/main/java/dk/sdu/mmmi/cbse/project1/common/Physics.java
@@ -2,8 +2,8 @@ package dk.sdu.mmmi.cbse.project1.common;
 
 public class Physics {
 
-	public Double drag = 1.0;
-	public Double velocityX = 0.0;
-	public Double velocityY = 0.0;
+	public float drag = 1.0f;
+	public float velocityX = 0.0f;
+	public float velocityY = 0.0f;
 
 }
